restructure the retry loop
this should make the send and check for the correct response more
stable. It will no longer kick out and delete stale packets, it will
find the correct ID packet if it is hidden deeper in the buffer. This
also makes it so the read is in only one mode and the packets are always
read in the same way.